requirements.txt
Django==4.0.3
dj-database-url==0.5.0
djangorestframework==3.12.4
mongoengine==0.24.0
python-decouple==3.4
pytz==2021.3
requests==2.26.0
